requirements.txt
jupyter_client==8.6.2
langchain==0.2.6
langchain_community==0.2.6
pandas==1.4.4
python-dotenv==1.0.1
streamlit==1.32.0
